# infra/openapi.yaml
...
components:
  schemas:
    Coverage:
      description: 'Hourly coverage of the returned window, so clients can render
        gaps.


        Computed over the requested window when ``from``/``to`` are supplied, else

        over the span of the returned rows. ``present_hours`` counts *distinct*

        hourly ``snapshot_at`` values (snapshots are top-of-hour UTC), so it is

        correct even for ``sid=null`` responses that carry several rows per hour.

        Both endpoints are inclusive. ``truncated`` is true when the row ``limit``

        bounded the result -- i.e. older data exists beyond ``window_start`` -- so
        a

        high ``missing_hours`` there reflects the cap, not gaps.'
      properties:
        expected_hours:
          title: Expected Hours
          type: integer
        missing_hours:
          title: Missing Hours
          type: integer
        present_hours:
          title: Present Hours
          type: integer
        truncated:
          title: Truncated
          type: boolean
        window_end:
          format: date-time
          title: Window End
          type: string
        window_start:
          format: date-time
          title: Window Start
          type: string
      required:
      - window_start
      - window_end
      - expected_hours
      - present_hours
      - missing_hours
      - truncated
      title: Coverage
      type: object
    DailyCoverage:
      description: 'Daily coverage of the returned window, so clients can render gaps.


        The daily analogue of :class:`Coverage`: ``present_days`` counts *distinct*

        ``trade_date`` values in the window (correct even for ``sid=null`` responses

        that carry several rows per day). Both endpoints are inclusive. ``truncated``

        is true when the row ``limit`` bounded the result.'
      properties:
        expected_days:
          title: Expected Days
          type: integer
        missing_days:
          title: Missing Days
          type: integer
        present_days:
          title: Present Days
          type: integer
        truncated:
          title: Truncated
          type: boolean
        window_end:
          format: date
          title: Window End
          type: string
        window_start:
          format: date
          title: Window Start
          type: string
      required:
      - window_start
      - window_end
      - expected_days
      - present_days
      - missing_days
      - truncated
      title: DailyCoverage
      type: object
    DailyResponse:
      description: Response body for ``GET /v1/market/items/{item_id}/daily`` (FR-14).
      properties:
        count:
          title: Count
          type: integer
        coverage:
          anyOf:
          - $ref: '#/components/schemas/DailyCoverage'
          - type: 'null'
        daily:
          items:
            $ref: '#/components/schemas/DailyRow'
          title: Daily
          type: array
        item_id:
          title: Item Id
          type: integer
        region:
          title: Region
          type: string
        sid:
          anyOf:
          - type: integer
          - type: 'null'
          title: Sid
      required:
      - item_id
      - region
      - sid
      - count
      - coverage
      - daily
      title: DailyResponse
      type: object
    DailyRow:
      description: market_daily table row.
      properties:
        avg_price:
          title: Avg Price
          type: integer
        avg_stock:
          title: Avg Stock
          type: integer
        close_price:
          title: Close Price
          type: integer
        high_price:
          title: High Price
          type: integer
        item_id:
          title: Item Id
          type: integer
        low_price:
          title: Low Price
          type: integer
        open_price:
          title: Open Price
          type: integer
        region:
          title: Region
          type: string
        sid:
          title: Sid
          type: integer
        snapshot_count:
          title: Snapshot Count
          type: integer
        total_trades_delta:
          title: Total Trades Delta
          type: integer
        trade_date:
          format: date
          title: Trade Date
          type: string
      required:
      - region
      - trade_date
      - item_id
      - sid
      - open_price
      - high_price
      - low_price
      - close_price
      - avg_price
      - total_trades_delta
      - avg_stock
      - snapshot_count
      title: DailyRow
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    SnapshotsResponse:
      description: Response body for ``GET /v1/market/items/{item_id}/snapshots``
        (FR-13).
      properties:
        count:
          title: Count
          type: integer
        coverage:
          anyOf:
          - $ref: '#/components/schemas/Coverage'
          - type: 'null'
        item_id:
          title: Item Id
          type: integer
        region:
          title: Region
          type: string
        sid:
          anyOf:
          - type: integer
          - type: 'null'
          title: Sid
        snapshots:
          items:
            $ref: '#/components/schemas/SnapshotRow'
          title: Snapshots
          type: array
      required:
      - item_id
      - region
      - sid
      - count
      - coverage
      - snapshots
      title: SnapshotsResponse
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: x-api-key
      type: apiKey
info:
  title: BDO Market Insights API
  version: 1.0.0
openapi: 3.1.0
paths:
  /v1/insights:
    get:
      operationId: get_insights_v1_insights_get
      parameters:
      - description: 'BDO server region. Accepted: na, eu, sea, mena, kr, ru, jp,
          th, tw, sa, console_eu, console_na, console_asia. Default: tw.'
        in: query
        name: region
        required: false
        schema:
          default: tw
          description: 'BDO server region. Accepted: na, eu, sea, mena, kr, ru, jp,
            th, tw, sa, console_eu, console_na, console_asia. Default: tw.'
          enum:
          - na
          - eu
          - sea
          - mena
          - kr
          - ru
          - jp
          - th
          - tw
          - sa
          - console_eu
          - console_na
          - console_asia
          title: Region
          type: string
      - description: 'Summary cadence. Accepted: daily, weekly. Default: daily.'
        in: query
        name: period
        required: false
        schema:
          default: daily
          description: 'Summary cadence. Accepted: daily, weekly. Default: daily.'
          enum:
          - daily
          - weekly
          title: Period
          type: string
      - description: 'Summary date, ISO-8601 (YYYY-MM-DD). Default: latest available.'
        in: query
        name: date
        required: false
        schema:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          description: 'Summary date, ISO-8601 (YYYY-MM-DD). Default: latest available.'
          title: Date
      - description: 'Narrative language code. Default: en.'
        in: query
        name: lang
        required: false
        schema:
          default: en
          description: 'Narrative language code. Default: en.'
          title: Lang
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                title: Return
                type: string
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: GET /v1/insights
...
  /v1/market/items/{item_id}/analysis:
    get:
      operationId: get_analysis_v1_market_items__item_id__analysis_get
      parameters:
      - in: path
        name: item_id
        required: true
        schema:
          title: Item Id
          type: integer
      - description: 'BDO server region. Accepted: na, eu, sea, mena, kr, ru, jp,
          th, tw, sa, console_eu, console_na, console_asia. Default: tw.'
        in: query
        name: region
        required: false
        schema:
          default: tw
          description: 'BDO server region. Accepted: na, eu, sea, mena, kr, ru, jp,
            th, tw, sa, console_eu, console_na, console_asia. Default: tw.'
          enum:
          - na
          - eu
          - sea
          - mena
          - kr
          - ru
          - jp
          - th
          - tw
          - sa
          - console_eu
          - console_na
          - console_asia
          title: Region
          type: string
      - description: 'Enhancement sub-id. Default: 0 (the base item).'
        in: query
        name: sid
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: 'Enhancement sub-id. Default: 0 (the base item).'
          title: Sid
      - description: 'Trailing analytics window in days. Range: 1-90. Default: 14.
          Needs at least 7 daily points.'
        in: query
        name: window_days
        required: false
        schema:
          default: 14
          description: 'Trailing analytics window in days. Range: 1-90. Default: 14.
            Needs at least 7 daily points.'
          maximum: 90.0
          minimum: 1.0
          title: Window Days
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Return
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: GET /v1/market/items/{item_id}/analysis
  /v1/market/items/{item_id}/daily:
    get:
      operationId: get_daily_v1_market_items__item_id__daily_get
      parameters:
      - in: path
        name: item_id
        required: true
        schema:
          title: Item Id
          type: integer
      - description: 'BDO server region. Accepted: na, eu, sea, mena, kr, ru, jp,
          th, tw, sa, console_eu, console_na, console_asia. Default: tw.'
        in: query
        name: region
        required: false
        schema:
          default: tw
          description: 'BDO server region. Accepted: na, eu, sea, mena, kr, ru, jp,
            th, tw, sa, console_eu, console_na, console_asia. Default: tw.'
          enum:
          - na
          - eu
          - sea
          - mena
          - kr
          - ru
          - jp
          - th
          - tw
          - sa
          - console_eu
          - console_na
          - console_asia
          title: Region
          type: string
      - description: 'Enhancement sub-id. Default: all sids.'
        in: query
        name: sid
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: 'Enhancement sub-id. Default: all sids.'
          title: Sid
      - description: 'Lower bound, inclusive, ISO-8601 date (YYYY-MM-DD). Default:
          90 days before today when neither from nor to is given.'
        in: query
        name: from
        required: false
        schema:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          description: 'Lower bound, inclusive, ISO-8601 date (YYYY-MM-DD). Default:
            90 days before today when neither from nor to is given.'
          title: From
      - description: 'Upper bound, inclusive, ISO-8601 date (YYYY-MM-DD). Default:
          unbounded.'
        in: query
        name: to
        required: false
        schema:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          description: 'Upper bound, inclusive, ISO-8601 date (YYYY-MM-DD). Default:
            unbounded.'
          title: To
      - description: 'Max daily rows returned (hard cap). Range: 1-990, clamped. The
          returned span is normally bounded by the default 90-day window rather than
          by limit.'
        in: query
        name: limit
        required: false
        schema:
          default: 990
          description: 'Max daily rows returned (hard cap). Range: 1-990, clamped.
            The returned span is normally bounded by the default 90-day window rather
            than by limit.'
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: GET /v1/market/items/{item_id}/daily
  /v1/market/items/{item_id}/snapshots:
    get:
      operationId: get_snapshots_v1_market_items__item_id__snapshots_get
      parameters:
      - in: path
        name: item_id
        required: true
        schema:
          title: Item Id
          type: integer
      - description: 'BDO server region. Accepted: na, eu, sea, mena, kr, ru, jp,
          th, tw, sa, console_eu, console_na, console_asia. Default: tw.'
        in: query
        name: region
        required: false
        schema:
          default: tw
          description: 'BDO server region. Accepted: na, eu, sea, mena, kr, ru, jp,
            th, tw, sa, console_eu, console_na, console_asia. Default: tw.'
          enum:
          - na
          - eu
          - sea
          - mena
          - kr
          - ru
          - jp
          - th
          - tw
          - sa
          - console_eu
          - console_na
          - console_asia
          title: Region
          type: string
      - description: 'Enhancement sub-id. Default: all sids.'
        in: query
        name: sid
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: 'Enhancement sub-id. Default: all sids.'
          title: Sid
      - description: 'Lower bound, inclusive, ISO-8601 datetime. Default: 168 hours
          before now when neither from nor to is given.'
        in: query
        name: from
        required: false
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          description: 'Lower bound, inclusive, ISO-8601 datetime. Default: 168 hours
            before now when neither from nor to is given.'
          title: From
      - description: 'Upper bound, inclusive, ISO-8601 datetime. Default: unbounded.'
        in: query
        name: to
        required: false
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          description: 'Upper bound, inclusive, ISO-8601 datetime. Default: unbounded.'
          title: To
      - description: 'Max snapshots returned (hard cap). Range: 1-1848, clamped. The
          returned span is normally bounded by the default 168-hour window rather
          than by limit; set it only to cap rows explicitly.'
        in: query
        name: limit
        required: false
        schema:
          default: 1848
          description: 'Max snapshots returned (hard cap). Range: 1-1848, clamped.
            The returned span is normally bounded by the default 168-hour window rather
            than by limit; set it only to cap rows explicitly.'
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotsResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: GET /v1/market/items/{item_id}/snapshots
security:
- ApiKeyAuth: []
servers:
- url: /